[WIP] Added setters/getters to Action objects.
The idea behind this PR is to add setters/getters to Action objects. All
the properties that belongs to the actions should be accesible directly
from the object, without any external calls to the store.

This PR adds a fallback mechanism reading most unset properties from the
store, but ideally the store should create the action object property.

To be able to read the claim ID off an action id, this PR introduces a
new abstract method to the Store abstract class
(`get_claim_id_for_action`).

This PR wants to fix #94, #95 and #96.