Add ActionScheduler_StoredAction
For instantiating actions that have been saved in the data store,
and therefore, have an ID, status and may have a claim ID, without
touching the existing ActionScheduler_Action class so that there is
no change to it (for backward compatibility). Ideally, that class
would be renamed something like ActionScheduler_UnsavedAction or
ActionScheduler_UnstoredAction, as it's more descriptive of the
differences but this helps disambiguate an instance of a stored
action and an unstored action without changing the public APIs
from previous versions.

Also update ActionScheduler_FinishedAction to extend this new class
instead of ActionScheduler_Action so that finished actions can also
be instantiated with an ID, status etc.

Part of #85