Cleanup stores in ActionScheduler_ListTable
Firstly, inject the store and logger classes through the constructor
rather than accessing them from the global stores, to help with testing.

Secondly, don't store them both in the one property, because they are
quite different data, and the logger isn't really a store so doesn't
belong in a property named 'stores'.

Finally, set $this->store before calling maybe_render_admin_notices()
so that it can use the built-in store instead of having to call
ActionScheduler::store() directly to access the store (and negating
the need for a $store/$stores property).