Instantiate trashed/cancelled actions
Using the new ActionScheduler_CanceledAction action type, which
is near identical to ActionScheduler_FinishedAction in behaviour,
but will always return a Null schedule to make sure it is never
rescheduled.

The other changes are that instantiated trashed actions will now
return a valid hook and args, which is helpful for #85, and shouldn't
cause issues with backward compatibility.

That said, this is a dangerous change, as evidenced by the changes
required to the unite tests, as this could lead to trashed/cancelled
actions being reschedule or run, but I can't see anyway that would
happen after reviewing the QueueRunner code, and also with the way
ActionScheduler_CanceledAction will always return a null schedule.