Add ActionScheduler_AbstractAction
For use as a shared base for both stored and unstored actions (i.e.
the original ActionScheduler_Action objects that do not have an ID,
status or accessible claim ID).

This has a number of flow on benefits, like being able to require
an ID and status in the constructors when instantiating stored actions,
which come in a few different instances.

It also means we need to deprecated ActionScheduler_FinishedAction,
because for those instances, we now need an instance of an object that
extends ActionScheduler_StoredAction so we can access the action's ID
and status, and to do that, we also want to require those values to be
passed into the objects constructor.

ActionScheduler_NonexecutableAction is introduced for that purpose and
to replace ActionScheduler_FinishedAction.

Having ActionScheduler_AbstractAction also allows for the null action,
ActionScheduler_NullAction, to extend it.