Add as_enqueue_async_action() and async APIs
For scheduling actions that will always be run as soon as possible,
rather than a date in the future, or even now, but only after other
scheduled actions with a date in the past.

This commit makes it acceptable to save an action with a NULL schedule
by mapping its MySQL datetime string to 0000-00-00 00:00:00. This
makes sure that any "async" action is given priority in queue processing.