Remove docs recommendation to load AS with negative priority
Hi,

I'm not sure if opening a GH issue first would be better, than editing the docs directly... but here I am :)

In our plugin, we were requiring the AS library via the `plugins_loaded` action with priority `-10`. This requiring/loading of AS is not working as expected. I think the loading goes into the "Theme usage" use case, if there are no other plugins registering AS and this caused a fatal PHP error once we tested it with WooCommerce 5.5 Nightly (Trunk) -> which has the newer AS 3.2.0 version in it (our plugin is using 3.1.6 for now).

The Fatal error was:

```
PHP Fatal error:  Cannot declare class ActionScheduler, because the name is already in use in /.../wp-content/plugins/woocommerce/packages/action-scheduler/classes/abstracts/ActionScheduler.php on line 10
```

A simple fix was to require the AS library as soon as our plugin initializes (before the `plugins_loaded` hook is fired). So that's why I think the removed part of this doc (this PR) is needed.