Fix: Use orderby => 'none' when bulk cancelling actions.
The orderby clause when querying actions to cancel can result in a filesort operation for larger tables as there is no corresponding index that provides the where clause followed by the action_id.  Since any actions being cancelled transition to another status, we don't need to have a sort in the initial query to read the IDs.