Fix possible error when option value is ''
Somehow (maybe old AS version?) the option value can get set to ''. In this case the code will attempt to insert the option but it will give a duplicate key error. So the lock can never be set until someone manually deletes this option. Now the option will be updated if it's '' so the code can work.